2021/НТО, 2-й этап/SOLVED chinasocialcredit/requirements.txt
Pillow==8.3.2
pygame==2.0.2
